Fix the plugin tests
